+++ src/Profile/Rule.cs
@@ -67,7 +67,7 @@ namespace KERBALISM
 			Resource_info res = input.Length > 0 ? resources.Info(v, input) : null;
 
 			// determine message variant
-			uint variant = vi.temperature < Settings.SurvivalTemperature ? 0 : 1u;
+			uint variant = vi.temperature < PreferencesLifeSupport.Instance.survivalTemperature ? 0 : 1u;
 
 			// get product of all environment modifiers
 			double k = Modifiers.Evaluate(v, vi, resources, modifiers);